requirements.txt
aerich==0.1.9
aiofiles==0.5.0
aiomysql==0.0.20
aiosqlite==0.13.0
alembic==1.4.2
anyio==1.3.0
async-generator==1.10
asyncclick==7.0.9
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
databases==0.3.2
dnspython==1.16.0
email-validator==1.1.1
Faker==4.1.0
fastapi==0.55.1
h11==0.9.0
idna==2.9
iso8601==0.1.12
jose==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
passlib==1.7.2
protobuf==3.12.2
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
PyMySQL==0.9.2
PyPika==0.37.6
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-multipart==0.0.5
qiniu==7.2.8
redis==3.5.3
requests==2.23.0
six==1.15.0
sniffio==1.1.0
SQLAlchemy==1.3.17
starlette==0.13.2
tencentcloud-sdk-python==3.0.182
text-unidecode==1.3
tortoise-orm==0.16.12
typer==0.2.1
typing-extensions==3.7.4.2
unicorn==1.0.1
urllib3==1.25.9
uvicorn==0.11.5
websockets==8.1
